+++ services/comments.service.js
@@ -1,11 +1,26 @@
 const CommentRepo = require('../repositories/comments.repository');
-const { MakeError } = require('../utils/makeErrorUtil');
+const MakeError = require('../utils/makeErrorUtil');
 
 class CommentService {
   commentRepo = new CommentRepo();
 
   createComment = async (userId, cardId, content) => {
     try {
+      if (!userId) {
+        throw new MakeError(400, 'userId를 수신받지 못했습니다.');
+      }
+      if (isNaN(userId) || userId < 1) {
+        throw new MakeError(400, 'userId가 유효한 형식이 아닙니다.');
+      }
+      if (!cardId) {
+        throw new MakeError(400, 'cardId를 수신받지 못했습니다.');
+      }
+      if (isNaN(cardId) || cardId < 1) {
+        throw new MakeError(400, 'cardId가 유효한 형식이 아닙니다.');
+      }
+      if (!content) {
+        throw new MakeError(400, '내용을 입력해 주세요.');
+      }
       const createCommentResult = await this.commentRepo.createComment(
         userId,
         cardId,
@@ -22,7 +37,13 @@ class CommentService {
 
   getAllComments = async (cardId) => {
     try {
-      const getCommentsResult = await this.cardRepo.getAllComment(cardId);
+      if (!cardId) {
+        throw new MakeError(400, 'cardId를 수신받지 못했습니다.');
+      }
+      if (isNaN(cardId) || cardId < 1) {
+        throw new MakeError(400, 'cardId가 유효한 형식이 아닙니다.');
+      }
+      const getCommentsResult = await this.commentRepo.getAllComments(cardId);
 
       if (getCommentsResult.length === 0) {
         throw new MakeError(400, '데이터가 존재하지 않습니다.');
@@ -30,13 +51,19 @@ class CommentService {
 
       return getCommentsResult;
     } catch (err) {
-      console.error('CommentService_CommentAllComments', error);
-      throw error;
+      console.error('CommentService_CommentAllComments', err);
+      throw err;
     }
   };
 
   getComment = async (commentId) => {
     try {
+      if (!commentId) {
+        throw new MakeError(400, 'commentId를 수신받지 못했습니다.');
+      }
+      if (isNaN(commentId) || commentId < 1) {
+        throw new MakeError(400, 'commentId가 올바른 형식이 아닙니다.');
+      }
       const getCommentResult = await this.commentRepo.getComment(commentId);
 
       if (!getCommentResult) {
@@ -45,13 +72,19 @@ class CommentService {
 
       return getCommentResult;
     } catch (err) {
-      console.error('CommentService_CommentComment', error);
-      throw error;
+      console.error('CommentService_CommentComment', err);
+      throw err;
     }
   };
 
   deleteComment = async (userId, commentId) => {
     try {
+      if (!commentId) {
+        throw MakeError(400, 'commentId를 수신받지 못했습니다.');
+      }
+      if (isNaN(commentId) || commentId < 1) {
+        throw new MakeError(400, 'commentId가 유효한 형식이 아닙니다.');
+      }
       const deleteCommentResult = await this.commentRepo.deleteComment(
         userId,
         commentId,
